+++ spg/source/utils/graphColoring.cpp
@@ -11,7 +11,7 @@ void Graph::setNumNodes(const int numNodes)
 }
 void Graph::addEdge(const int i, const int j)
 {
-    if (i >= m_adjacency.size() || j >= m_adjacency.size()) {
+    if (i >= static_cast<int>(m_adjacency.size()) || j >= static_cast<int>(m_adjacency.size())) {
         throw std::runtime_error("Trying to add edge to graph with invalid node indices");
     }
     m_adjacency[i].push_back(j);
@@ -52,7 +52,7 @@ std::vector<int> colorStencils(const FlatStencils &flatStencils)
 {
     Graph graph;
     std::vector<std::list<int>> stencilsPerVertex;
-    for (int i = 0; i < flatStencils.entries.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(flatStencils.entries.size()); ++i) {
         const int vertexId = flatStencils.entries[i];
         const int stencilId = i / flatStencils.stencilSize;
         stencilsPerVertex.resize(std::max(stencilsPerVertex.size(), static_cast<size_t>(vertexId + 1)));
@@ -74,7 +74,7 @@ std::vector<int> colorVertices(int numVertices, const std::vector<FlatStencils> 
     Graph graph;
     graph.setNumNodes(numVertices);
     for (auto &flatStencils : flatStencilsSet) {
-        for (int i = 0; i < flatStencils.entries.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(flatStencils.entries.size()); ++i) {
             for (int j = i + 1; j % flatStencils.stencilSize != 0; ++j) {
                 graph.addEdge(flatStencils.entries[i], flatStencils.entries[j]);
             }
